requirements.txt
aiobotocore==0.9.4
aiodns==1.1.1
aiofiles==0.4.0
aiohttp==3.4.4
asn1crypto==0.24.0
async-timeout==3.0.0
attrs==18.2.0
backcall==0.1.0
bcrypt==3.1.4
bleach==3.0.2
boto3==1.9.0
botocore==1.10.58
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
coincurve==9.0.0
coloredlogs==10.0
colorlog==3.1.4
cytoolz==0.9.0.1
decorator==4.3.0
defusedxml==0.5.0
diffiehellman==0.13.3
docutils==0.14
eciespy==0.1.2
entrypoints==0.2.3
eth-hash==0.2.0
eth-keys==0.2.0b3
eth-typing==1.3.0
eth-utils==1.2.1
Faker==0.9.0
grpcio==1.14.2
grpcio-tools==1.14.2
httptools==0.0.11
humanfriendly==4.16.1
idna==2.7
idna-ssl==1.1.0
ipykernel==5.1.0
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==0.24
jedi==0.12.1
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
MarkupSafe==1.0
mistune==0.8.4
multidict==4.4.0
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.0
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
prometheus-client==0.4.1
prompt-toolkit==1.0.15
protobuf==3.6.1
ptyprocess==0.6.0
pycares==2.3.0
pycparser==2.18
pycryptodome==3.6.6
pycryptodomex==3.6.6
Pygments==2.2.0
pygogo==0.12.0
pymongo==3.7.1
pysha3==1.0.2
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.2
requests==2.19.1
rethinkdb==2.3.0.post6
s3transfer==0.1.13
sanic==0.7.0
sawtooth-sdk==1.0.5
sawtooth-signing==1.0.5
secp256k1==0.13.2
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
terminado==0.8.1
testpath==0.4.2
text-unidecode==1.2
toml==0.9.4
toolz==0.9.0
tornado==5.1.1
traitlets==4.3.2
ujson==1.35
urllib3==1.23
uvloop==0.11.2
wcwidth==0.1.7
webencodings==0.5.1
websockets==6.0
widgetsnbextension==3.4.2
wrapt==1.10.11
yarl==1.2.6
